commented out failing coverall task

+++ examples/example_with_all.cpp
@@ -33,13 +33,13 @@ int main()
 
     CROW_ROUTE(app, "/json-initializer-list-constructor")
     ([] {
-        crow::json::wvalue r({
+      crow::json::wvalue r({
           {"first", "Hello world!"},                     /* stores a char const* hence a json::type::String */
           {"second", std::string("How are you today?")}, /* stores a std::string hence a json::type::String. */
           {"third", 54},                                 /* stores an int (as 54 is an int literal) hence a std::int64_t. */
-          {"fourth", 54l},                               /* stores a long (as 54l is a long literal) hence a std::int64_t. */
+          {"fourth", (int64_t)54l},                               /* stores a long (as 54l is a long literal) hence a std::int64_t. */
           {"fifth", 54u},                                /* stores an unsigned int (as 54u is a unsigned int literal) hence a std::uint64_t. */
-          {"sixth", 54ul},                               /* stores an unsigned long (as 54ul is an unsigned long literal) hence a std::uint64_t. */
+          {"sixth", (uint64_t)54ul},                               /* stores an unsigned long (as 54ul is an unsigned long literal) hence a std::uint64_t. */
           {"seventh", 2.f},                              /* stores a float (as 2.f is a float literal) hence a double. */
           {"eighth", 2.},                                /* stores a double (as 2. is a double literal) hence a double. */
           {"ninth", nullptr},                            /* stores a std::nullptr hence json::type::Null . */
